requirements.txt
ftfy
regex
tqdm
lucid
blobfile
tensorflow-gpu==1.13.2